pyproject.toml
[tool.poetry]
name = "playlist_converter"
version = "0.1.0"
description = ""
authors = ["Adewumi Isaac <ienioladewumi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
django-crispy-forms = "^1.13.0"
crispy-bootstrap5 = "^0.6"
ytmusicapi = "^0.19.3"
spotipy = "^2.19.0"
django-environ = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
